requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
APScheduler==3.11.0
asyncio==4.0.0
asyncpg==0.30.0
attrs==25.4.0
certifi==2025.10.5
et_xmlfile==2.0.0
frozenlist==1.8.0
greenlet==3.2.4
idna==3.10
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.3
propcache==0.4.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.23
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
yarl==1.22.0
